requirements.txt
streamlit
langchain
pillow